requirements.txt
django
djangorestframework
django-filter
drf-yasg
djangorestframework-simplejwt
mysqlclient
